// shared/schema.d.ts
export declare const auditLogSchema: z.ZodObject<{
    user_id: z.ZodOptional<z.ZodNumber>;
    action: z.ZodString;
    channel: z.ZodOptional<z.ZodEnum<["email", "sms"]>>;
    purpose: z.ZodOptional<z.ZodEnum<["signup", "login", "reset"]>>;
    status: z.ZodEnum<["success", "failure", "blocked"]>;
    ip_address: z.ZodOptional<z.ZodString>;
    user_agent: z.ZodOptional<z.ZodString>;
    details: z.ZodOptional<z.ZodString>;
}, "strip", z.ZodTypeAny, {
    action: string;
    status: "success" | "failure" | "blocked";
    user_id?: number | undefined;
    channel?: "email" | "sms" | undefined;
    purpose?: "signup" | "login" | "reset" | undefined;
    ip_address?: string | undefined;
    user_agent?: string | undefined;
    details?: string | undefined;
}, {
    action: string;
    status: "success" | "failure" | "blocked";
    user_id?: number | undefined;
    channel?: "email" | "sms" | undefined;
    purpose?: "signup" | "login" | "reset" | undefined;
    ip_address?: string | undefined;
    user_agent?: string | undefined;
    details?: string | undefined;
}>;